Toolkit: Add UT CMake macros with Catch2
Catch2 is the choosen UT framework this season. The UT.cmake toolkit
adds the add_ut macro, mostly copied from last year, for adding unit
tests, using the Catch2 framework. It'll only be enabled with the
UNITTESTS_ENABLED=ON CMake option.

If Valgrind is available, the toolkit will also generate a target
for executing UTs under Valgrind.

The target for running all non-valgrind UTs in a build is "check".
With the "check_valg" target, all UTs will be executed under Valgrind
to identify memory and such.