Toolkit: Add CMake preamble with common settings
The preamble.cmake file contains some common settings for this year's
Advent of Code. It sets the required C++ standard to C++17, enables
the export of compile commands for IDE usage and includes the UT.cmake
UT toolkit.

Furthermore, -Wall, -Werror and -Wextra are added to the compile flags
when the build type is Debug.