Common: Implement AoC Main Function
The AoC main function takes argc and argv as inputs and executes the
AoCDay class provided as a template argument.

The program needs to be provided with exactly 2 inputs, the part of the
day's puzzle to execute ("1" or "2") and the path to the input file.

The Day's class is instantiated, as well as the Tokenizer it requires.
The input file is then handed off to the Tokenizer.

Note that the aoc_main does not execute the tokenize(const char)
function - that is left to the Day class to allow for using different
delimiters on the same input for the two different parts.

With invalid inputs, a AoCError is thrown.