Common: Add CMake template func for Day build
The init_aoc_day function can be used to generate the build system
for a new AoC puzzle day.

Note: The function assumes that it is generating a build system for
a standalone CMake project.

Furthermore, please read the sources to see which files are assumed
to exist in the source directory.